refactor footer into RSC
