+++ src/data/footer_links.ts
@@ -1,0 +1,35 @@
+const FOOTER_LINKS = {
+  about_links: [
+    { name: 'What we do', url: '/what-we-do/' },
+    { name: 'Governance', url: '/governance/' },
+    { name: 'Steering and Working Groups', url: '/#' },
+    { name: 'Team', url: '/team/' },
+    { name: 'Job Opportunities', url: '/job-opportunities/' },
+    { name: 'Projects', url: '/projects/' }
+  ],
+  services_links: [
+    { name: 'Create DOIs', url: '/create-dois/' },
+    { name: 'Integrate Workflows', url: '/integrate-workflows/' },
+    { name: 'Enable Discovery', url: '/enable-discovery/' },
+    { name: 'Promote Reuse', url: '/promote-reuse/' },
+    { name: 'Strategic Initiatives', url: '/#' }
+  ],
+  resources_links: [
+    { name: 'Metadata Schema', url: 'https://schema.datacite.org' },
+    { name: 'Support', url: 'https://support.datacite.org' }
+  ],
+  community_links: [
+    { name: 'Become a Member', url: '/become-a-member/' },
+    { name: 'DataCite Fee Model', url: '/fee-model/' },
+    { name: 'Membership Enquiry', url: '/membership-enquiry/' },
+    { name: 'DataCite Members', url: '/members/' }
+  ],
+  contact_links: [
+    { name: 'Privacy Policy', url: '/privacy-policy/' },
+    { name: 'Terms and Conditions', url: '/terms-and-conditions/' },
+    { name: 'Imprint', url: '/imprint/' }
+
+  ]
+}
+
+export default FOOTER_LINKS
